% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdarima.R
\name{ltdarima_estimation}
\alias{ltdarima_estimation}
\title{Title}
\usage{
ltdarima_estimation(
  data,
  mean = FALSE,
  X = NULL,
  regular,
  seasonal,
  fixed_phi = TRUE,
  fixed_bphi = TRUE,
  fixed_theta = FALSE,
  fixed_btheta = FALSE,
  fixed_var = TRUE,
  eps = 1e-07,
  parametrization = c("mean_delta", "start_end")
)
}
\arguments{
\item{eps}{}
}
\description{
Title
}
